nitpick changes and move utils
