+++ packages/carbon/src/abstracts/BaseComponentInteraction.ts
@@ -7,11 +7,10 @@ import {
 } from "discord-api-types/v10"
 import type { Client } from "../classes/Client.js"
 import type { MessagePayload } from "../types/index.js"
-import { serializePayload, splitCustomId } from "../utils.js"
+import { serializePayload } from "../utils/index.js"
 import { BaseInteraction, type InteractionDefaults } from "./BaseInteraction.js"
 
 export class BaseComponentInteraction extends BaseInteraction<APIMessageComponentInteraction> {
-	customId: string
 	componentType: ComponentType
 	constructor(
 		client: Client,
@@ -21,7 +20,6 @@ export class BaseComponentInteraction extends BaseInteraction<APIMessageComponen
 		super(client, data, defaults)
 		if (!data.data)
 			throw new Error("Invalid interaction data was used to create this class")
-		this.customId = splitCustomId(data.data.custom_id)[0]
 		this.componentType = data.data.component_type
 	}
