+++ packages/carbon/src/abstracts/BaseGuildChannel.ts
@@ -9,8 +9,8 @@ import {
 import { Guild } from "../structures/Guild.js"
 import type { GuildCategoryChannel } from "../structures/GuildCategoryChannel.js"
 import type { MessagePayload } from "../types/index.js"
-import type { IfPartial } from "../utils.js"
-import { serializePayload } from "../utils.js"
+import type { IfPartial } from "../types/index.js"
+import { serializePayload } from "../utils/index.js"
 import { BaseChannel } from "./BaseChannel.js"
 
 export abstract class BaseGuildChannel<
