+++ packages/carbon/src/abstracts/BaseGuildTextChannel.ts
@@ -8,7 +8,7 @@ import {
 } from "discord-api-types/v10"
 import { GuildThreadChannel } from "../structures/GuildThreadChannel.js"
 import { Message } from "../structures/Message.js"
-import type { IfPartial } from "../utils.js"
+import type { IfPartial } from "../types/index.js"
 import { BaseGuildChannel } from "./BaseGuildChannel.js"
 
 export abstract class BaseGuildTextChannel<
