+++ packages/carbon/src/classes/Modal.ts
@@ -7,6 +7,7 @@ import type { BaseModalComponent } from "../abstracts/BaseModalComponent.js"
 import type { ModalInteraction } from "../internals/ModalInteraction.js"
 import type { ComponentData, ComponentParserResult } from "../types/index.js"
 import type { Row } from "./components/Row.js"
+import { parseCustomId } from "../utils/customIdParser.js"
 
 export abstract class Modal {
 	/**
@@ -39,21 +40,7 @@ export abstract class Modal {
 	 * @param id - The custom ID of the component as received from an interaction event
 	 * @returns The base key and the data object
 	 */
-	customIdParser: (id: string) => ComponentParserResult = (id) => {
-		const [key, ...data] = id.split(":")
-		if (!key) throw new Error(`Invalid component ID: ${id}`)
-		return {
-			key,
-			data: Object.fromEntries(
-				data.map((d) => {
-					const [k, v] = d.split("=")
-					if (v === "true") return [k, true]
-					if (v === "false") return [k, false]
-					return [k, Number.isNaN(Number(v)) ? v : Number(v)]
-				})
-			)
-		}
-	}
+	customIdParser: (id: string) => ComponentParserResult = parseCustomId
 
 	abstract run(
 		interaction: ModalInteraction,
