+++ packages/carbon/src/index.ts
@@ -85,4 +85,4 @@ export * from "./structures/User.js"
 export * from "discord-api-types/v10"
 export * from "./types/index.js"
 export * from "./permissions.js"
-export * from "./utils.js"
+export * from "./utils/index.js"
