+++ packages/carbon/src/structures/Guild.ts
@@ -10,7 +10,7 @@ import { Base } from "../abstracts/Base.js"
 import type { Client } from "../classes/Client.js"
 import { DiscordError } from "../errors/DiscordError.js"
 import { channelFactory } from "../functions/channelFactory.js"
-import type { IfPartial } from "../utils.js"
+import type { IfPartial } from "../types/index.js"
 import { GuildMember } from "./GuildMember.js"
 import { Role } from "./Role.js"
