+++ packages/carbon/src/structures/GuildMember.ts
@@ -2,7 +2,7 @@ import type { APIGuildMember, GuildMemberFlags } from "discord-api-types/v10"
 import { Base } from "../abstracts/Base.js"
 import type { Client } from "../classes/Client.js"
 import { maxPermissions } from "../permissions.js"
-import type { IfPartial } from "../utils.js"
+import type { IfPartial } from "../types/index.js"
 import type { Guild } from "./Guild.js"
 import { Role } from "./Role.js"
 import { User } from "./User.js"
