+++ packages/carbon/src/structures/GuildStageOrVoiceChannel.ts
@@ -5,7 +5,7 @@ import {
 	VideoQualityMode
 } from "discord-api-types/v10"
 import { BaseGuildChannel } from "../abstracts/BaseGuildChannel.js"
-import type { IfPartial } from "../utils.js"
+import type { IfPartial } from "../types/index.js"
 
 export abstract class GuildStageOrVoiceChannel<
 	Type extends ChannelType.GuildStageVoice | ChannelType.GuildVoice,
