+++ packages/carbon/src/structures/Message.ts
@@ -21,8 +21,8 @@ import { Base } from "../abstracts/Base.js"
 import type { Client } from "../classes/Client.js"
 import { Embed } from "../classes/Embed.js"
 import { channelFactory } from "../functions/channelFactory.js"
-import type { MessagePayload } from "../types/index.js"
-import { type IfPartial, serializePayload } from "../utils.js"
+import type { IfPartial, MessagePayload } from "../types/index.js"
+import { serializePayload } from "../utils/index.js"
 import { GuildThreadChannel } from "./GuildThreadChannel.js"
 import { Role } from "./Role.js"
 import { User } from "./User.js"
