+++ packages/carbon/src/types/index.ts
@@ -83,5 +83,6 @@ export type MessagePayloadFile = {
 }
 
 export type ArrayOrSingle<T> = T | T[]
+export type IfPartial<T, U, V = U | undefined> = T extends true ? V : U
 
 export * from "./listeners.js"
