+++ website/content/classes/modals.mdx
@@ -9,7 +9,16 @@ Modals are popup forms that can be used to collect user input. They are created 
 Modals are automatically registered when they are shown to a user. The registration happens when you call `showModal` on an interaction.
 
 ```ts title="src/commands/modal.ts"
-import { Command, type CommandInteraction, Modal, type ModalInteraction, Row, TextInput, TextInputStyle } from "@buape/carbon"
+import { 
+	Command, 
+	type CommandInteraction, 
+	type ComponentData, 
+	type Modal, 
+	type ModalInteraction, 
+	Row, 
+	TextInput, 
+	TextInputStyle
+} from "@buape/carbon"
 
 export default class ModalCommand extends Command {
 	name = "modal"
@@ -29,7 +38,7 @@ class TestModal extends Modal {
 		new Row([new TextInputAge()])
 	]
 
-	async run(interaction: ModalInteraction) {
+	async run(interaction: ModalInteraction, data: ComponentData) {
 		const name = interaction.fields.getText("name")
 		const age = interaction.fields.getText("age")
 		await interaction.reply(
